refactor: feature folder naming convention
