+++ src/features/games-list/utils/handleRefreshGamesList.ts
