+++ src/features/games-list/utils/handleSortingChange.ts
