+++ src/shared/components/layouts/Dashboard.tsx
@@ -3,8 +3,8 @@ import { useEffect } from 'react'
 import { Achievements } from '@/features/achievement-manager'
 import { AchievementOrderPage, AchievementUnlocker } from '@/features/achievement-unlocker'
 import { CardFarming } from '@/features/card-farming'
-import { CustomList } from '@/features/customlists'
-import { FreeGamesList, GamesList, IdlingGamesList } from '@/features/gameslist'
+import { CustomList } from '@/features/custom-lists'
+import { FreeGamesList, GamesList, IdlingGamesList } from '@/features/games-list'
 import { TradingCardsList } from '@/features/inventory-manager'
 import { Settings } from '@/features/settings'
 import { Sidebar } from '@/shared/components'
