+++ src/shared/hooks/useFreeGames.ts
@@ -1,5 +1,5 @@
 import { useCallback, useEffect, useRef } from 'react'
-import { useGamesList } from '@/features/gameslist'
+import { useGamesList } from '@/features/games-list'
 import { useUserStore } from '@/shared/stores'
 import { autoRedeemFreeGames, checkForFreeGames, hasGamerFeature } from '@/shared/utils'
